+++ src/main/java/com/avancial/app/data/databean/importMotrice/ImportTMDETVOEntity.java
@@ -16,13 +16,13 @@ public class ImportTMDETVOEntity {
 
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    @Id
-   private Long    idTMDETVO;
+   private Double    idTMDETVO;
    private String  ETVO_VOIT_COD_CIE;
    private String  ETVO_VOIT_NUM_TRA1;
    private String  ETVO_VOIT_IND_FER;
-   private Long    ETVO_VOIT_TRCH_NUM;
-   private Long    ETVO_VOIT_NUM;
-   private Long    ETVO_NUM;
+   private Double    ETVO_VOIT_TRCH_NUM;
+   private Double    ETVO_VOIT_NUM;
+   private Double    ETVO_NUM;
    private String  ETVO_ETAT_VOIT;
    private Boolean ETVO_REGI;
    private String  ETVO_USER;
@@ -31,7 +31,7 @@ public class ImportTMDETVOEntity {
    /**
     * @return the idTMDETVO
     */
-   public Long getIdTMDETVO() {
+   public Double getIdTMDETVO() {
       return this.idTMDETVO;
    }
 
@@ -39,7 +39,7 @@ public class ImportTMDETVOEntity {
     * @param idTMDETVO
     *           the idTMDETVO to set
     */
-   public void setIdTMDETVO(Long idTMDETVO) {
+   public void setIdTMDETVO(Double idTMDETVO) {
       this.idTMDETVO = idTMDETVO;
    }
 
@@ -91,7 +91,7 @@ public class ImportTMDETVOEntity {
    /**
     * @return the eTVO_VOIT_TRCH_NUM
     */
-   public Long getETVO_VOIT_TRCH_NUM() {
+   public Double getETVO_VOIT_TRCH_NUM() {
       return this.ETVO_VOIT_TRCH_NUM;
    }
 
@@ -99,14 +99,14 @@ public class ImportTMDETVOEntity {
     * @param eTVO_VOIT_TRCH_NUM
     *           the eTVO_VOIT_TRCH_NUM to set
     */
-   public void setETVO_VOIT_TRCH_NUM(Long eTVO_VOIT_TRCH_NUM) {
+   public void setETVO_VOIT_TRCH_NUM(Double eTVO_VOIT_TRCH_NUM) {
       this.ETVO_VOIT_TRCH_NUM = eTVO_VOIT_TRCH_NUM;
    }
 
    /**
     * @return the eTVO_VOIT_NUM
     */
-   public Long getETVO_VOIT_NUM() {
+   public Double getETVO_VOIT_NUM() {
       return this.ETVO_VOIT_NUM;
    }
 
@@ -114,14 +114,14 @@ public class ImportTMDETVOEntity {
     * @param eTVO_VOIT_NUM
     *           the eTVO_VOIT_NUM to set
     */
-   public void setETVO_VOIT_NUM(Long eTVO_VOIT_NUM) {
+   public void setETVO_VOIT_NUM(Double eTVO_VOIT_NUM) {
       this.ETVO_VOIT_NUM = eTVO_VOIT_NUM;
    }
 
    /**
     * @return the eTVO_NUM
     */
-   public Long getETVO_NUM() {
+   public Double getETVO_NUM() {
       return this.ETVO_NUM;
    }
 
@@ -129,7 +129,7 @@ public class ImportTMDETVOEntity {
     * @param eTVO_NUM
     *           the eTVO_NUM to set
     */
-   public void setETVO_NUM(Long eTVO_NUM) {
+   public void setETVO_NUM(Double eTVO_NUM) {
       this.ETVO_NUM = eTVO_NUM;
    }
